e1000: fix descriptor overflow
Allow rxq->rx_tail + offset > 65535
in eth_em_rx_descriptor_done().

+++ lib/librte_pmd_e1000/em_rxtx.c
@@ -1449,7 +1449,7 @@ eth_em_rx_descriptor_done(void *rx_queue, uint16_t offset)
 {
 	volatile struct e1000_rx_desc *rxdp;
 	struct em_rx_queue *rxq = rx_queue;
-	uint16_t desc;
+	uint32_t desc;
 
 	if (unlikely(offset >= rxq->nb_rx_desc))
 		return 0;
